requirements.txt
qdrant-client
streamlit
answer-generation
sentence-transformers
transformers
torch
openai-whisper
